fix(client): remove legacy handshake fallback and fix top_k type
The native client previously treated an unexpected or absent
`HelloAckFrame` magic as a signal to silently continue with default
negotiated values, hiding protocol mismatches. Both paths now return a
structured `NodeDbError::internal` so callers can diagnose the failure.

`top_k` in the vector search request is corrected from `u64` to `u32` to
match the protocol field type. The graph traversal depth helper is
similarly corrected from `Option<u64>` to `Option<u32>`, and the
vector delete plan builder adds an explicit `u32` bounds check on
`vector_id` with a `BadRequest` error on overflow.

+++ nodedb-client/src/native/client.rs
@@ -157,7 +157,7 @@ impl NodeDb for NativeClient {
                 TextFields {
                     collection: Some(collection.to_string()),
                     query_vector: Some(query.to_vec()),
-                    top_k: Some(k as u64),
+                    top_k: Some(k as u32),
                     ..Default::default()
                 },
             )
